modify controller
