+++ app/assets/javascripts/genres.coffee
