+++ app/assets/javascripts/user_products.coffee
