+++ app/assets/stylesheets/disks.scss
@@ -1,3 +1,3 @@
-// Place all the styles related to the admin_disks controller here.
+// Place all the styles related to the disks controller here.
 // They will automatically be included in application.css.
 // You can use Sass (SCSS) here: http://sass-lang.com/
