+++ app/assets/stylesheets/genres.scss
@@ -1,3 +1,3 @@
-// Place all the styles related to the admin_genres controller here.
+// Place all the styles related to the genres controller here.
 // They will automatically be included in application.css.
 // You can use Sass (SCSS) here: http://sass-lang.com/
