+++ app/assets/stylesheets/products.scss
@@ -1,3 +1,3 @@
-// Place all the styles related to the admin_products controller here.
+// Place all the styles related to the products controller here.
 // They will automatically be included in application.css.
 // You can use Sass (SCSS) here: http://sass-lang.com/
