+++ app/assets/stylesheets/song_orders.scss
@@ -1,3 +1,3 @@
-// Place all the styles related to the admin_songs_order controller here.
+// Place all the styles related to the song_orders controller here.
 // They will automatically be included in application.css.
 // You can use Sass (SCSS) here: http://sass-lang.com/
