--- app/controllers/ user_products_controller.rb
@@ -1,7 +1,0 @@
-class ProductsController < ApplicationController
-  def index
-  end
-
-  def show
-  end
-end
