+++ app/controllers/genres_controller.rb
@@ -1,13 +1,13 @@
-class AdminGenresController < ApplicationController
+class GenresController < ApplicationController
   def index
+  end
+
+  def edit
   end
 
   def create
   end
 
   def update
-  end
-
-  def edit
   end
 end
