+++ app/controllers/products_controller.rb
@@ -1,17 +1,17 @@
-class AdminProductsController < ApplicationController
+class ProductsController < ApplicationController
   def new
-  end
-
-  def create
-  end
-
-  def update
   end
 
   def index
   end
 
   def edit
+  end
+
+  def create
+  end
+
+  def update
   end
 
   def destroy
