+++ app/controllers/songs_controller.rb
@@ -1,10 +1,10 @@
-class AdminSongsController < ApplicationController
+class SongsController < ApplicationController
+  def new
+  end
   def create
   end
 
   def update
   end
 
-  def new
-  end
 end
