+++ app/controllers/user_products_controller.rb
@@ -1,0 +1,7 @@
+class UserProductsController < ApplicationController
+  def index
+  end
+
+  def show
+  end
+end
