--- app/helpers/admin_disks_helper.rb
@@ -1,2 +1,0 @@
-module AdminDisksHelper
-end
