--- app/helpers/admin_genres_helper.rb
@@ -1,2 +1,0 @@
-module AdminGenresHelper
-end
