--- app/helpers/admin_products_helper.rb
@@ -1,2 +1,0 @@
-module AdminProductsHelper
-end
