--- app/helpers/admin_songs_helper.rb
@@ -1,2 +1,0 @@
-module AdminSongsHelper
-end
