--- app/helpers/admin_songs_order_helper.rb
@@ -1,2 +1,0 @@
-module AdminSongsOrderHelper
-end
