+++ app/helpers/disks_helper.rb
@@ -1,0 +1,2 @@
+module DisksHelper
+end
