+++ app/helpers/genres_helper.rb
@@ -1,0 +1,2 @@
+module GenresHelper
+end
