+++ app/helpers/song_orders_helper.rb
@@ -1,0 +1,2 @@
+module SongOrdersHelper
+end
