+++ app/helpers/songs_helper.rb
@@ -1,0 +1,2 @@
+module SongsHelper
+end
