+++ app/helpers/user_products_helper.rb
@@ -1,0 +1,2 @@
+module UserProductsHelper
+end
