--- app/views/disks/create.html.erb
@@ -1,2 +1,0 @@
-<h1>AdminDisks#create</h1>
-<p>Find me in app/views/admin_disks/create.html.erb</p>
