--- app/views/disks/update.html.erb
@@ -1,2 +1,0 @@
-<h1>AdminDisks#update</h1>
-<p>Find me in app/views/admin_disks/update.html.erb</p>
