--- app/views/genres/create.html.erb
@@ -1,2 +1,0 @@
-<h1>AdminGenres#create</h1>
-<p>Find me in app/views/admin_genres/create.html.erb</p>
