+++ app/views/genres/edit.html.erb
@@ -1,2 +1,2 @@
-<h1>AdminGenres#edit</h1>
-<p>Find me in app/views/admin_genres/edit.html.erb</p>
+<h1>Genres#edit</h1>
+<p>Find me in app/views/genres/edit.html.erb</p>
