+++ app/views/genres/index.html.erb
@@ -1,2 +1,2 @@
-<h1>AdminGenres#index</h1>
-<p>Find me in app/views/admin_genres/index.html.erb</p>
+<h1>Genres#index</h1>
+<p>Find me in app/views/genres/index.html.erb</p>
