--- app/views/genres/update.html.erb
@@ -1,2 +1,0 @@
-<h1>AdminGenres#update</h1>
-<p>Find me in app/views/admin_genres/update.html.erb</p>
