--- app/views/products/create.html.erb
@@ -1,2 +1,0 @@
-<h1>AdminProducts#create</h1>
-<p>Find me in app/views/admin_products/create.html.erb</p>
