+++ app/views/products/edit.html.erb
@@ -1,2 +1,2 @@
-<h1>AdminProducts#edit</h1>
-<p>Find me in app/views/admin_products/edit.html.erb</p>
+<h1>Products#edit</h1>
+<p>Find me in app/views/products/edit.html.erb</p>
