+++ app/views/products/index.html.erb
@@ -1,2 +1,2 @@
-<h1>AdminProducts#index</h1>
-<p>Find me in app/views/admin_products/index.html.erb</p>
+<h1>Products#index</h1>
+<p>Find me in app/views/products/index.html.erb</p>
