+++ app/views/products/new.html.erb
@@ -1,2 +1,2 @@
-<h1>AdminProducts#new</h1>
-<p>Find me in app/views/admin_products/new.html.erb</p>
+<h1>Products#new</h1>
+<p>Find me in app/views/products/new.html.erb</p>
