--- app/views/products/update.html.erb
@@ -1,2 +1,0 @@
-<h1>AdminProducts#update</h1>
-<p>Find me in app/views/admin_products/update.html.erb</p>
