--- app/views/song_orders/create.html.erb
@@ -1,2 +1,0 @@
-<h1>AdminSongsOrder#create</h1>
-<p>Find me in app/views/admin_songs_order/create.html.erb</p>
