--- app/views/song_orders/update.html.erb
@@ -1,2 +1,0 @@
-<h1>AdminSongsOrder#update</h1>
-<p>Find me in app/views/admin_songs_order/update.html.erb</p>
