--- app/views/songs/create.html.erb
@@ -1,2 +1,0 @@
-<h1>AdminSongs#create</h1>
-<p>Find me in app/views/admin_songs/create.html.erb</p>
