+++ app/views/songs/new.html.erb
@@ -1,0 +1,2 @@
+<h1>Songs#new</h1>
+<p>Find me in app/views/songs/new.html.erb</p>
