--- app/views/songs/update.html.erb
@@ -1,2 +1,0 @@
-<h1>AdminSongs#update</h1>
-<p>Find me in app/views/admin_songs/update.html.erb</p>
