+++ app/views/user_products/index.html.erb
@@ -1,2 +1,2 @@
-<h1>Products#index</h1>
-<p>Find me in app/views/products/index.html.erb</p>
+<h1>UserProducts#index</h1>
+<p>Find me in app/views/user_products/index.html.erb</p>
