+++ app/views/user_products/show.html.erb
@@ -1,2 +1,2 @@
-<h1>Products#show</h1>
-<p>Find me in app/views/products/show.html.erb</p>
+<h1>UserProducts#show</h1>
+<p>Find me in app/views/user_products/show.html.erb</p>
