--- test/controllers/admin_genres_controller_test.rb
@@ -1,24 +1,0 @@
-require 'test_helper'
-
-class AdminGenresControllerTest < ActionDispatch::IntegrationTest
-  test "should get index" do
-    get admin_genres_index_url
-    assert_response :success
-  end
-
-  test "should get create" do
-    get admin_genres_create_url
-    assert_response :success
-  end
-
-  test "should get update" do
-    get admin_genres_update_url
-    assert_response :success
-  end
-
-  test "should get edit" do
-    get admin_genres_edit_url
-    assert_response :success
-  end
-
-end
