--- test/controllers/admin_products_controller_test.rb
@@ -1,29 +1,0 @@
-require 'test_helper'
-
-class AdminProductsControllerTest < ActionDispatch::IntegrationTest
-  test "should get new" do
-    get admin_products_new_url
-    assert_response :success
-  end
-
-  test "should get create" do
-    get admin_products_create_url
-    assert_response :success
-  end
-
-  test "should get update" do
-    get admin_products_update_url
-    assert_response :success
-  end
-
-  test "should get index" do
-    get admin_products_index_url
-    assert_response :success
-  end
-
-  test "should get edit" do
-    get admin_products_edit_url
-    assert_response :success
-  end
-
-end
