+++ test/controllers/products_controller_test.rb
@@ -1,13 +1,18 @@
 require 'test_helper'
 
 class ProductsControllerTest < ActionDispatch::IntegrationTest
+  test "should get new" do
+    get products_new_url
+    assert_response :success
+  end
+
   test "should get index" do
     get products_index_url
     assert_response :success
   end
 
-  test "should get show" do
-    get products_show_url
+  test "should get edit" do
+    get products_edit_url
     assert_response :success
   end
