+++ test/controllers/song_orders_controller_test.rb
@@ -1,0 +1,7 @@
+require 'test_helper'
+
+class SongOrdersControllerTest < ActionDispatch::IntegrationTest
+  # test "the truth" do
+  #   assert true
+  # end
+end
