+++ test/controllers/songs_controller_test.rb
@@ -1,0 +1,9 @@
+require 'test_helper'
+
+class SongsControllerTest < ActionDispatch::IntegrationTest
+  test "should get new" do
+    get songs_new_url
+    assert_response :success
+  end
+
+end
