+++ test/controllers/user_products_controller_test.rb
@@ -1,0 +1,14 @@
+require 'test_helper'
+
+class UserProductsControllerTest < ActionDispatch::IntegrationTest
+  test "should get index" do
+    get user_products_index_url
+    assert_response :success
+  end
+
+  test "should get show" do
+    get user_products_show_url
+    assert_response :success
+  end
+
+end
